pyproject.toml
[tool.poetry]
name = "capital-copilot"
version = "0.1.0"
description = "Capital Copilot is a tool to help aggregate market data, setup alerts, automate trading, and provide AI insights."
authors = ["Nik <nik.cimino@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = "^5.0.3"
django-allauth = "^0.61.1"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
openbb = {extras = ["alpha-vantage", "charting"], version = "^4.1.6"}
#openbb = {extras = ["all"], version = "^4.1.5"}
copilot-plugin-view = {path = "../copilot-plugin-view", develop = true}
copilot-plugin-dashboard = {path = "../copilot-plugin-dashboard", develop = true}

[tool.poetry.group.dev.dependencies]
langchain = "^0.1.11"
langchain-openai = "^0.0.8"
gitpython = "^3.1.42"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.9.0"
django-extensions = "^3.2.3"
pre-commit = "^3.6.2"
toml = "^0.10.2"
types-toml = "^0.10.8.20240310"
types-pyyaml = "^6.0.12.20240311"
django-stubs = "^4.2.7"
mdformat = "^0.7.17"
mdformat-black = "^0.1.1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
